pyproject.toml
[project]
name = "medicaltracker"
version = "0.1.0"
description = ""
authors = [
    {name = "damir",email = "1damiraminov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.0.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "pytz>=2024.1",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
